requirements.txt
aiogram
rich
GitPython
meval
fakeredis
redis